requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
asgiref=3.5.2=py39hecd8cb5_0
ca-certificates=2022.12.7=h033912b_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h6c40b1e_3
cryptography=38.0.4=py39hf6deb26_0
django=4.1=py39hecd8cb5_0
django-extensions=3.2.1=pyhd8ed1ab_0
django-rest-knox=4.2.0=pypi_0
django-rest-passwordreset=1.3.0=pypi_0
djangorestframework=3.14.0=pyhd8ed1ab_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.2=hecd8cb5_6
ncurses=6.3=hca72f7f_3
openssl=1.1.1s=hfd90126_1
pip=22.3.1=py39hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.6.0=pyhd8ed1ab_0
python=3.9.15=h218abb5_2
pytz=2022.7.1=pyhd8ed1ab_0
readline=8.2=hca72f7f_0
setuptools=65.6.3=py39hecd8cb5_0
sqlite=3.40.1=h880c91c_0
sqlparse=0.4.3=py39hecd8cb5_0
tk=8.6.12=h5d9f67b_0
tzdata=2022g=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
